Prevent compiler warnings (well, most of).
Unfortunately, code smell mostly comes from third-party sources,
like Jasper and Unrar, which may be re-imported again in the future.

Warnings that were left:

* Use of `tmpnam` in Jasper.
* Possible wrapping of `localtime_r` for multi-threaded wxWidgets.

+++ plugins/chartdldr_pi/src/unrar/extract.cpp
@@ -323,7 +323,7 @@ bool CmdExtract::ExtractCurrentFile(Archive &Arc,size_t HeaderSize,bool &Repeat)
       if (Cmd->VersionControl==0)
         MatchFound=false;
       int Version=ParseVersionFileName(ArcFileName,false);
-      if (Cmd->VersionControl-1==Version)
+      if (Cmd->VersionControl-1==static_cast<uint>(Version))
         ParseVersionFileName(ArcFileName,true);
       else
         MatchFound=false;
